+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/application/controller/OrdemServicoGestaoController.java
@@ -1,7 +1,10 @@
 package com.automotiva.estetica.rick.application.controller;
 
-import com.automotiva.estetica.rick.application.dto.request.*;
-import com.automotiva.estetica.rick.application.dto.response.AgendamentosHojeListResponse;
+import com.automotiva.estetica.rick.application.dto.request.AdicionarServicosOrdemRequest;
+import com.automotiva.estetica.rick.application.dto.request.AtualizarOrdemServicoGestaoRequest;
+import com.automotiva.estetica.rick.application.dto.request.AtualizarValorServicoOrdemRequest;
+import com.automotiva.estetica.rick.application.dto.request.OrdemServicoGestaoPageRequest;
+import com.automotiva.estetica.rick.application.dto.request.OrdemServicoRequest;
 import com.automotiva.estetica.rick.application.dto.response.OrdemServicoDetalheResponse;
 import com.automotiva.estetica.rick.application.dto.response.OrdemServicoResumoResponse;
 import com.automotiva.estetica.rick.application.security.GerenteOnly;
@@ -45,12 +48,6 @@ public class OrdemServicoGestaoController {
         return ResponseEntity.ok(ordemServicoUseCase.buscarTodosParaGestao(request));
     }
 
-    @GetMapping("/hoje")
-    @Operation(summary = "Lista todos os agendamentos do dia atual com detalhes completos para visualização no dashboard")
-    public ResponseEntity<AgendamentosHojeListResponse> buscarAgendamentosHoje() {
-        return ResponseEntity.ok(ordemServicoUseCase.buscarAgendamentosHoje());
-    }
-
     @GetMapping("/{ordemServicoId}")
     @Operation(summary = "Busca detalhes de uma ordem de serviço para gestão")
     public ResponseEntity<OrdemServicoDetalheResponse> buscarDetalheParaGestao(@PathVariable Long ordemServicoId) {
@@ -92,12 +89,5 @@ public class OrdemServicoGestaoController {
     public ResponseEntity<OrdemServicoDetalheResponse> removerServicoParaGestao(@PathVariable Long ordemServicoId,
             @PathVariable Long servicoId) {
         return ResponseEntity.ok(ordemServicoUseCase.removerServicoParaGestao(ordemServicoId, servicoId));
-    }
-
-    @PatchMapping("/{ordemServicoId}/cancel")
-    @Operation(summary = "Cancela uma ordem de serviço com motivo")
-    public ResponseEntity<OrdemServicoDetalheResponse> cancelarParaGestao(@PathVariable Long ordemServicoId,
-            @Valid @RequestBody CancelarOrdemRequest request) {
-        return ResponseEntity.ok(ordemServicoUseCase.cancelarParaGestao(ordemServicoId, request));
     }
 }
