+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/domain/usecase/PeriodoMensal.java
@@ -6,10 +6,8 @@ import java.time.LocalDateTime;
 public record PeriodoMensal(LocalDateTime inicio, LocalDateTime fim) {
 
     public static PeriodoMensal mesAtual() {
-        LocalDate hoje = LocalDate.now();
-        LocalDate inicio = hoje.withDayOfMonth(1);
-        LocalDate fim = hoje.withDayOfMonth(hoje.lengthOfMonth());
-        return new PeriodoMensal(inicio.atStartOfDay(), fim.atTime(23, 59, 59));
+        LocalDate inicio = LocalDate.now().withDayOfMonth(1);
+        return new PeriodoMensal(inicio.atStartOfDay(), LocalDateTime.now());
     }
 
     public static PeriodoMensal mesAnterior() {
