+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/domain/usecase/PeriodoMensalFactory.java
@@ -9,15 +9,13 @@ public final class PeriodoMensalFactory {
     }
 
     public static PeriodoMensal mesAtual() {
-        return PeriodoMensal.mesAtual();
-    }
-
-    public static PeriodoMensal ultimos30Dias() {
-        LocalDate inicio = LocalDate.now().minusDays(30);
+        LocalDate inicio = LocalDate.now().withDayOfMonth(1);
         return new PeriodoMensal(inicio.atStartOfDay(), LocalDateTime.now());
     }
 
     public static PeriodoMensal mesAnterior() {
-        return PeriodoMensal.mesAnterior();
+        LocalDate inicio = LocalDate.now().minusMonths(1).withDayOfMonth(1);
+        LocalDate fim = inicio.plusMonths(1).minusDays(1);
+        return new PeriodoMensal(inicio.atStartOfDay(), fim.atTime(23, 59, 59));
     }
 }
